Using 'set -e' rather than '/bin/bash -e'
Though command in script was failing, it was not halting and was not returning error.
As an experiment, A script having a command which fails was writtent. After the failing command,
an echo statement was placed.
Script containing "#!/bin/bash -e" printed the echo
While script with "set -e" did not print the echo
Scripts was run using /bin/bash test.sh

p.s: Running script with ./test.sh showed the expected behavior in both cases.

Signed-off-by: tahir24434 <tahir.rauf1@gmail.com>